+++ civis/tests/mocks.py
@@ -76,6 +76,7 @@ def create_client_mock_for_container_tests(
     c.scripts.post_containers_runs.return_value = mock_container_run_start
     c.scripts.get_containers_runs.return_value = mock_container_run
     c.scripts.list_containers_runs_outputs.return_value = (run_outputs or [])
+    c.jobs.list_runs_outputs.return_value = (run_outputs or [])
     c.jobs.list_runs_logs.return_value = (log_outputs or [])
 
     def change_state_to_cancelled(script_id):